Move alla input-types to their own folders
