+++ src/Form/types/Label/index.js
@@ -1,0 +1,3 @@
+import Label from './Label';
+
+export default Label;
