+++ src/Form/types/Text/index.js
@@ -1,0 +1,3 @@
+import Text from './Text';
+
+export default Text;
