+++ src/Form/types/TextArea/index.js
@@ -1,0 +1,3 @@
+import TextArea from './TextArea';
+
+export default TextArea;
